Allow fractional ppm correction
0.1ppm should be enough for anybody... O:-)

But seriously, both my hackrf and rtlsdr are something like X.4-X.6 ppm and
I really want my waterfall display to line up with my reference frequency.

+++ applications/gqrx/receiver.cpp
@@ -538,7 +538,7 @@ receiver::status receiver::set_filter_shape(filter_shape shape)
 }
 **/
 
-receiver::status receiver::set_freq_corr(int ppm)
+receiver::status receiver::set_freq_corr(double ppm)
 {
     src->set_freq_corr(ppm);
